Fix: Upload working again | Work: Details TBA
Bumpted Version of Multer to ^1.0.1 to minimize bugs caused by it.
Bumpted Version of Async to 1.4 o minimize bugs caused by it and the new
forEachOf method.

The way fileuploads got handled seemed wrong and got changed to the way it
is shown in some angluar samples. The Server handling has to be modified
to represent the frontend changes.

This changes will work but need a review and stylguide/quality check by
one of the original programmers.